Keeper endpoint testing
